fix implementation of logger
